Address security settings review
