+++ Meshtastic/Enums/KeyBackupStatus.swift
@@ -4,27 +4,6 @@
 //
 
 import Foundation
-import CryptoKit
-
-struct IdentityKeyPairBackup: Codable, Equatable {
-	let privateKey: Data
-	let publicKey: Data
-
-	static func isValid(privateKey: Data, publicKey: Data) -> Bool {
-		guard privateKey.count == 32, publicKey.count == 32 else {
-			return false
-		}
-
-		do {
-			let derivedPublicKey = try Curve25519.KeyAgreement.PrivateKey(
-				rawRepresentation: privateKey
-			).publicKey.rawRepresentation
-			return derivedPublicKey == publicKey
-		} catch {
-			return false
-		}
-	}
-}
 
 enum KeyBackupStatus: Equatable {
 	case saved
