+++ Meshtastic/Resources/docs/index.json
@@ -398,35 +398,35 @@
             "app",
             "radio",
             "node",
+            "device",
             "from",
             "firmware",
-            "device",
             "traffic",
-            "nodes",
-            "mesh",
             "settings",
+            "nodes",
+            "only",
+            "mesh",
             "position",
             "tag",
             "region",
             "language",
             "nfc",
-            "only",
+            "packet",
+            "default",
             "connected",
-            "configure",
+            "available",
             "translation",
             "rate",
             "packets",
-            "packet",
-            "default",
+            "key",
+            "identity",
             "contact",
+            "configure",
             "channel",
             "translated",
-            "setting",
-            "set",
-            "messages",
-            "management"
+            "setting"
         ],
-        "charCount": 12440
+        "charCount": 13175
     },
     {
         "id": "signal-meter",
